Move HighLine spec helpers to support
Moves the main support methods for testing against HighLine into the
spec/support/ui.rb for reuse in other classes.  The `shared_context` can
be included by adding `:with_ui` to the current `describe`/`context`
block.

Functionality should remain the same, but some other expectation helpers
were added as well:

* expect_readline_question_asked

  This tests specifically that a chunk of text has been output through
  readline.  Since the `expect_heard` only will check against the first
  chunk of text passed to that method from readline, this allows you be
  more specific about which text is tested against.

  Used by `expect_heard` under the covers

* expect_output

  This specifically tests against all of the output that wasn't pushed
  through `readline` only.

  Used by `expect_heard` under the covers
