+++ spec/prompts_spec.rb
@@ -1,43 +1,8 @@
-require 'readline'
+require 'support/ui'
 
-describe ManageIQ::ApplianceConsole::Prompts do
-  let(:input) do
-    @temp_stdin = Tempfile.new("temp_stdin")
-    File.open(@temp_stdin.path, 'w+')
-  end
-  let(:readline_output) do
-    @temp_stdout = Tempfile.new("temp_stdout")
-    File.open(@temp_stdout.path, 'w+')
-  end
-  let(:output) { StringIO.new }
-  let(:prompt) { "\n?  " }
-  after do
-    @temp_stdin.close! if @temp_stdin
-    @temp_stdout.close! if @temp_stdout
-  end
-
+describe ManageIQ::ApplianceConsole::Prompts, :with_ui do
   subject do
-    Readline.input = input
-    Readline.output = readline_output
     Class.new(HighLine) { include ManageIQ::ApplianceConsole::Prompts }.new(input, output)
-  end
-
-  after do
-    # best-guess cleanup: Readline has .input=, .output= but no .input, .output
-    Readline.input = STDIN
-    Readline.output = STDOUT
-  end
-
-  # net/ssh messes with track_eof
-  # we need it for testing
-  # set it back after we are done
-
-  before do
-    @old_track, HighLine.track_eof = HighLine.track_eof?, true
-  end
-
-  after do
-    HighLine.track_eof = @old_track
   end
 
   context "#sample_url" do
@@ -579,29 +544,5 @@ describe ManageIQ::ApplianceConsole::Prompts do
       expect(subject.just_ask("prompt", "default")).to eq("this")
       expect_heard("Enter the prompt: |default| ")
     end
-  end
-
-  private
-
-  def say(strs)
-    Array(strs).each do |str|
-      input << str << "\n"
-    end
-    input.rewind
-  end
-
-  def expect_cls
-    expect(subject).to receive(:print)
-  end
-
-  def expect_heard(strs, check_eof = true)
-    readline_output.rewind
-    readline_output_content = readline_output.read
-    strs = Array(strs)
-    expect(readline_output_content).to include(strs.shift) unless readline_output_content.empty?
-    strs = strs.collect { |s| s == "" ? "\n" : s }.join
-    expect(output.string).to eq(strs)
-    expect { subject.ask("is there more") }.to raise_error(EOFError) if check_eof
-    expect(input).to be_eof
   end
 end
